pyproject.toml
[project]
name = "echoes-of-deceit"
version = "0.1.0"
description = "Turtle Soup puzzle game with AI agents"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    # ==================== LangChain ====================
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-ollama>=0.3.0",
    "langchain-openai>=0.3.0",
    
    # ==================== LangGraph ====================
    "langgraph>=0.2.0",
    "langgraph-checkpoint>=2.0.0",
    
    # ==================== RAG ====================
    "lightrag-hku[api]>=1.4.8",
    
    # ==================== Web Framework ====================
    "streamlit>=1.29.0",
    
    # ==================== HTTP & Async ====================
    "httpx",
    "nest-asyncio>=1.5.0",
    
    # ==================== Data & Config ====================
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "isort",
    "mypy",
]

[project.scripts]
echoes = "src.cli:main"
play = "src.play:main"

# ==================== Git Dependencies ====================
[tool.uv.sources]
minirag = { git = "https://github.com/Linorman/MiniRAG.git" }

# ==================== UV Configuration ====================
[tool.uv]
# Development mode dependencies
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ==================== Tool Configurations ====================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
